feat: add Nightwatch for end-to-end testing (#257)

+++ index.ts
@@ -14,6 +14,7 @@ import { postOrderDirectoryTraverse, preOrderDirectoryTraverse } from './utils/d
 import generateReadme from './utils/generateReadme'
 import getCommand from './utils/getCommand'
 import renderEslint from './utils/renderEslint'
+import { FILES_TO_FILTER } from './utils/filterList'
 
 function isValidPackageName(projectName) {
   return /^(?:@[a-z0-9-*~][a-z0-9-*._~]*\/)?[a-z0-9-~][a-z0-9-._~]*$/.test(projectName)
@@ -75,6 +76,7 @@ async function init() {
   // --with-tests / --tests (equals to `--vitest --cypress`)
   // --vitest
   // --cypress
+  // --nightwatch
   // --playwright
   // --eslint
   // --eslint-with-prettier (only support prettier through eslint for simplicity)
@@ -101,6 +103,7 @@ async function init() {
       argv.tests ??
       argv.vitest ??
       argv.cypress ??
+      argv.nightwatch ??
       argv.playwright ??
       argv.eslint
     ) === 'boolean'
@@ -119,7 +122,7 @@ async function init() {
     needsRouter?: boolean
     needsPinia?: boolean
     needsVitest?: boolean
-    needsE2eTesting?: false | 'cypress' | 'playwright'
+    needsE2eTesting?: false | 'cypress' | 'nightwatch' | 'playwright'
     needsEslint?: boolean
     needsPrettier?: boolean
   } = {}
@@ -134,6 +137,7 @@ async function init() {
     // - Install Vue Router for SPA development?
     // - Install Pinia for state management?
     // - Add Cypress for testing?
+    // - Add Nightwatch for testing?
     // - Add Playwright for end-to-end testing?
     // - Add ESLint for code quality?
     // - Add Prettier for code formatting?
@@ -227,6 +231,13 @@ async function init() {
               value: 'cypress'
             },
             {
+              title: 'Nightwatch',
+              description: answers.needsVitest
+                ? undefined
+                : 'also supports unit testing with Nightwatch Component Testing',
+              value: 'nightwatch'
+            },
+            {
               title: 'Playwright',
               value: 'playwright'
             }
@@ -283,6 +294,8 @@ async function init() {
   const { needsE2eTesting } = result
   const needsCypress = argv.cypress || argv.tests || needsE2eTesting === 'cypress'
   const needsCypressCT = needsCypress && !needsVitest
+  const needsNightwatch = argv.nightwatch || argv.tests || needsE2eTesting === 'nightwatch'
+  const needsNightwatchCT = needsNightwatch && !needsVitest
   const needsPlaywright = argv.playwright || needsE2eTesting === 'playwright'
 
   const root = path.join(cwd, targetDir)
@@ -329,6 +342,12 @@ async function init() {
   if (needsCypressCT) {
     render('config/cypress-ct')
   }
+  if (needsNightwatch) {
+    render('config/nightwatch')
+  }
+  if (needsNightwatchCT) {
+    render('config/nightwatch-ct')
+  }
   if (needsPlaywright) {
     render('config/playwright')
   }
@@ -348,6 +367,12 @@ async function init() {
     }
     if (needsVitest) {
       render('tsconfig/vitest')
+    }
+    if (needsNightwatch) {
+      render('tsconfig/nightwatch')
+    }
+    if (needsNightwatchCT) {
+      render('tsconfig/nightwatch-ct')
     }
   }
 
@@ -393,7 +418,7 @@ async function init() {
       root,
       () => {},
       (filepath) => {
-        if (filepath.endsWith('.js')) {
+        if (filepath.endsWith('.js') && !FILES_TO_FILTER.includes(path.basename(filepath))) {
           const tsFilePath = filepath.replace(/\.js$/, '.ts')
           if (fs.existsSync(tsFilePath)) {
             fs.unlinkSync(filepath)
@@ -437,7 +462,9 @@ async function init() {
       needsTypeScript,
       needsVitest,
       needsCypress,
+      needsNightwatch,
       needsPlaywright,
+      needsNightwatchCT,
       needsCypressCT,
       needsEslint
     })
